Revise
Note to self: This revision came after the writer made a realization that not everyone has the bandwidth or interest to read an online post as though it’s a book.

Keep in mind:
Value people’s time
Get to the point
Stop being a smart-ass
Stop expounding pessimism

In this revision, I removed sections that were unhelpfully distracting from the post’s ideas, passages that may entertain the patient reader but are off-putting to the many who are just browsing the internet. In effect, I uncovered the essence by clearing up the dust. Remember: be clear and guide the reader.